Port upstream changes onto the AutoPtr/MonoPtr object lifetime system:
- TextureLoadOptions/SRGBMode, KTX2 loading, TextureCacheKey-keyed cache
- DescriptorTableManager thread safety and slot refcounting
- SceneTypeFactory::CreateGraph / Scene::CreateSceneGraph
- Vulkan update-after-bind check, ray tracing validation/position fetch
- ManipulatorPass (rebuilds the pick view since PlanarView is non-copyable)
- ARM64 lzcnt fix, Agility SDK 720 test guard

imgui_nvrhi.cpp stays removed; its degenerate clip-rect fix is ported to
ImGuiRenderPass.cpp. LoadTextureFromRawImageMemory now takes
TextureLoadOptions instead of bool sRGB.

+++ src/app/ImGuiRenderPass.cpp
@@ -455,7 +455,8 @@ void ImGui_ImplNVRHI_RenderDrawData(ImDrawData *draw_data, nvrhi::IFramebuffer *
                                 (draw_cmd->ClipRect.y - clip_off.y) * clip_scale.y);
                 ImVec2 clip_max((draw_cmd->ClipRect.z - clip_off.x) * clip_scale.x,
                                 (draw_cmd->ClipRect.w - clip_off.y) * clip_scale.y);
-                if (clip_max.x <= clip_min.x || clip_max.y <= clip_min.y) continue;
+                // Test the integer rect: a sub-pixel clip truncates to an empty scissor.
+                if ((int)clip_max.x <= (int)clip_min.x || (int)clip_max.y <= (int)clip_min.y) continue;
 
                 // Apply scissor/clipping rectangle
                 state.viewport.scissorRects[0] = nvrhi::Rect{
